+++ Passepartout/Sources/AppConstants.swift
@@ -197,13 +197,16 @@ public class AppConstants {
 
         public static let referrals: [Infrastructure.Name: String] = [
             .mullvad: "https://mullvad.net/en/account/create/",
+            .nordVPN: "https://join.nordvpn.com/order/",
             .pia: "https://www.privateinternetaccess.com/pages/buy-vpn/",
             .protonVPN: "https://protonvpn.net/?aid=keeshux",
             .tunnelBear: "https://click.tunnelbear.com/aff_c?offer_id=2&aff_id=7464",
             .windscribe: "https://secure.link/kCsD0prd"
         ]
 
-        public static let externalResources: [Infrastructure.Name: String] = [:]
+        public static let externalResources: [Infrastructure.Name: String] = [
+            .nordVPN: "https://downloads.nordcdn.com/configs/archives/certificates/servers.zip" // 9MB
+        ]
     }
 
     public class Repos {
@@ -252,6 +255,11 @@ public class AppConstants {
                 "PIATunnel",
                 "MIT",
                 "https://raw.githubusercontent.com/pia-foss/tunnel-apple/master/LICENSE"
+            ),
+            License(
+                "SSZipArchive",
+                "MIT",
+                "https://raw.githubusercontent.com/samsoffes/ssziparchive/master/LICENSE"
             ),
             License(
                 "SwiftGen",