+++ Passepartout/Sources/Services/Infrastructure.swift
@@ -24,12 +24,14 @@
 //
 
 import Foundation
-import TunnelKit
+import SSZipArchive
 
 public struct Infrastructure: Codable {
     public enum Name: String, Codable, Comparable {
         case mullvad = "Mullvad"
         
+        case nordVPN = "NordVPN"
+
         case pia = "PIA"
         
         case protonVPN = "ProtonVPN"
@@ -62,10 +64,18 @@ public struct Infrastructure: Codable {
         return try JSONDecoder().decode(Infrastructure.self, from: json)
     }
     
+    public func defaultPool() -> Pool? {
+        return pool(withPrefix: defaults.pool)
+    }
+    
     public func pool(for identifier: String) -> Pool? {
         return pools.first { $0.id == identifier }
     }
 
+    public func pool(withPrefix prefix: String) -> Pool? {
+        return pools.first { $0.id.hasPrefix(prefix) }
+    }
+    
     public func preset(for identifier: String) -> InfrastructurePreset? {
         return presets.first { $0.id == identifier }
     }
@@ -87,10 +97,17 @@ extension Infrastructure.Name {
     }
 
     public func importExternalResources(from url: URL, completionHandler: @escaping () -> Void) {
+        var task: () -> Void
         switch self {
+        case .nordVPN:
+            task = {
+                SSZipArchive.unzipFile(atPath: url.path, toDestination: self.externalURL.path)
+            }
+            
         default:
-            break
+            task = {}
         }
+        execute(task: task, completionHandler: completionHandler)
     }
 
     private func execute(task: @escaping () -> Void, completionHandler: @escaping () -> Void) {